Fix database deadlock and improve structure
1. Keep a persistent database connection in PreferenceStore
2. Move class of core outside
3. Use a global coroutines scope in VectorDaemon
